+++ src/device/virtio/block/device.rs
@@ -22,7 +22,7 @@ use crate::device::virtio::features::{
     VIRTIO_F_IN_ORDER, VIRTIO_F_RING_EVENT_IDX, VIRTIO_F_VERSION_1,
 };
 use crate::device::virtio::{MmioConfig, SingleFdSignalQueue, QUEUE_MAX_SIZE};
-use crate::kvm::Hypervisor;
+use crate::kvm::hypervisor::Hypervisor;
 
 use super::inorder_handler::InOrderQueueHandler;
 use super::queue_handler::QueueHandler;
