+++ src/inspect.rs
@@ -12,7 +12,7 @@ pub struct InspectOptions {
 
 pub fn inspect(opts: &InspectOptions) -> Result<()> {
     let vm = try_with!(
-        kvm::get_hypervisor(opts.pid),
+        kvm::hypervisor::get_hypervisor(opts.pid),
         "cannot get vms for process {}",
         opts.pid
     );
